+++ Changelog.md
@@ -3,6 +3,7 @@
 ## v0.13.0 (unreleased)
 
 - Add an `EarlyListener` that allows sending of 0.5-RTT data.
+- Add a `TokenStore` to store address validation tokens.
 
 ## v0.12.0 (2019-08-05)